requirements.txt
Django==5.2.11
django-phonenumber-field
phonenumbers
python-decouple
gunicorn
whitenoise